Update 110622
Added quite a lot of things, like particules to the weapons, and tried to make the weapon spawn all over the place but fail. I was doing the weapon information but it seems that it will be hearder. Tomorrow have to do a play with what i have and move on.