Update 11092022
Changed the battery sprite for the laserGun prefab. Also changed the spawn point to only give +1 battery and move the UI to account for mobile devices. Got a new design for the UI > Health bar and energy that will make it a lot cleaner and new ideas for the batter and game loop.